refactor: restructure backend modules
