+++ core/src/backend/local/mod.rs
@@ -1,5 +1,4 @@
 mod core_backend;
-pub use core_backend::CoreBackend;
 mod directory;
 mod log;
 mod note;
