+++ core/src/backend/local/note.rs
