+++ core/src/schema.rs
@@ -1,7 +1,7 @@
 use {
     crate::{
         Error, Result,
-        db::{Execute, Storage},
+        backend::local::{Execute, Storage},
         types::DirectoryId,
     },
     gluesql::core::ast_builder::{col, table, text},
