+++ core/src/state/notebook/inner_state.rs
@@ -5,7 +5,7 @@ mod note_tree;
 
 use crate::{
     Event, NotebookEvent, NotebookTransition, Result,
-    db::CoreBackend,
+    backend::CoreBackend,
     state::notebook::{NotebookState, note},
     types::KeymapGroup,
 };
