+++ core/src/state/notebook/inner_state/editing_visual_mode.rs
@@ -1,6 +1,6 @@
 use crate::{
     Event, Result,
-    db::CoreBackend,
+    backend::CoreBackend,
     state::notebook::NotebookState,
     transition::{NotebookTransition, VisualModeTransition},
     types::KeymapGroup,
