+++ core/src/state/notebook/inner_state/editing_visual_mode/gateway.rs
@@ -1,6 +1,6 @@
 use crate::{
     Error, Event, KeyEvent, Result,
-    db::CoreBackend,
+    backend::CoreBackend,
     state::notebook::{InnerState, NotebookState, VimNormalState},
     transition::{NotebookTransition, VisualModeTransition},
     types::{KeymapGroup, KeymapItem},
