+++ core/src/state/notebook/inner_state/editing_visual_mode/numbering.rs
@@ -1,6 +1,6 @@
 use crate::{
     Error, Event, KeyEvent, Result,
-    db::CoreBackend,
+    backend::CoreBackend,
     state::notebook::{InnerState, NotebookState, VimNormalState},
     transition::{NormalModeTransition, NotebookTransition, VimKeymapKind, VisualModeTransition},
     types::{KeymapGroup, KeymapItem},
