+++ core/src/state/notebook/inner_state/note_tree/note_more_actions.rs
@@ -1,6 +1,6 @@
 use crate::{
     Error, Event, NotebookEvent, NotebookTransition, Result,
-    db::CoreBackend,
+    backend::CoreBackend,
     state::notebook::{NotebookState, note},
     types::{KeymapGroup, KeymapItem},
 };
